controller/requirements.txt
matplotlib
numpy
opencv-python
pyserial
pyzbar
